@@ -1,101 +1,109 @@
 {
-	"name": "@xcandy/ui-taro",
-	"version": "0.0.0-development",
-	"main": "dist/index.js",
-	"description": "A UI framework for taro.js",
-	"scripts": {
-		"release": "bash scripts/release.sh",
-		"test": "echo \"no test\"",
-		"lint": "npm run eslint",
-		"lint:fix": "npm run eslint:fix",
-		"eslint": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
-		"eslint:fix": "eslint --fix \"src/**/*.{js,jsx,ts,tsx}\"",
-		"ui:build": "cross-env TARO_BUILD_TYPE=ui taro build --ui",
-		"ui:ci": "TARO_BUILD_TYPE=ui taro build --ui",
-		"build:weapp": "taro build --type weapp",
-		"build:swan": "taro build --type swan",
-		"build:alipay": "taro build --type alipay",
-		"build:tt": "taro build --type tt",
-		"build:h5": "taro build --type h5",
-		"build:rn": "taro build --type rn",
-		"dev:weapp": "npm run build:weapp -- --watch",
-		"dev:swan": "npm run build:swan -- --watch",
-		"dev:alipay": "npm run build:alipay -- --watch",
-		"dev:tt": "npm run build:tt -- --watch",
-		"dev:h5": "npm run build:h5 -- --watch",
-		"dev:rn": "npm run build:rn -- --watch",
-		"update:site": "scripts/updateSite"
-	},
-	"author": "love@onelong.org",
-	"license": "MIT",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/xxxcandy/xcandy-ui-taro"
-	},
-	"keywords": [
-		"taro",
-		"react",
-		"ui framework",
-		"weapp"
-	],
-	"files": [
-		"dist",
-		"@types",
-		"src"
-	],
-	"types": "./@types/index.d.ts",
-	"husky": {
-		"hooks": {
-			"pre-commit": "npm run lint && lint-staged",
-			"commit-msg": "commitlint -e $GIT_PARAMS"
-		}
-	},
-	"lint-staged": {
-		"*.{js,jsx,ts,tsx}": [
-			"eslint --quiet --cache --fix",
-			"git add"
-		]
-	},
-	"dependencies": {
-		"@tarojs/components": "1.2.22",
-		"@tarojs/router": "1.2.22",
-		"@tarojs/taro": "1.2.22",
-		"@tarojs/taro-alipay": "1.2.22",
-		"@tarojs/taro-h5": "1.2.22",
-		"@tarojs/taro-swan": "1.2.22",
-		"@tarojs/taro-tt": "1.2.22",
-		"@tarojs/taro-weapp": "1.2.22",
-		"classnames": "^2.2.6",
-		"dayjs": "^1.8.0",
-		"nerv-devtools": "^1.3.9",
-		"nervjs": "^1.3.9"
-	},
-	"devDependencies": {
-		"@tarojs/plugin-babel": "1.2.22",
-		"@tarojs/plugin-csso": "1.2.22",
-		"@tarojs/plugin-sass": "1.2.22",
-		"@tarojs/plugin-uglifyjs": "1.2.22",
-		"@tarojs/webpack-runner": "1.2.22",
-		"@types/react": "^16.4.6",
-		"@types/webpack-env": "^1.13.6",
-		"babel-eslint": "^8.2.3",
-		"babel-plugin-transform-class-properties": "^6.24.1",
-		"babel-plugin-transform-decorators-legacy": "^1.3.4",
-		"babel-plugin-transform-jsx-stylesheet": "^0.6.5",
-		"babel-plugin-transform-object-rest-spread": "^6.26.0",
-		"babel-preset-env": "^1.6.1",
-		"commitlint": "^7.2.1",
-		"cross-env": "^5.2.0",
-		"eslint": "^4.15.0",
-		"eslint-config-taro": "1.2.22",
-		"eslint-plugin-import": "^2.12.0",
-		"eslint-plugin-jest": "^21.21.0",
-		"eslint-plugin-react": "^7.8.2",
-		"eslint-plugin-taro": "1.2.22",
-		"eslint-plugin-typescript": "^0.14.0",
-		"husky": "^1.3.1",
-		"lint-staged": "^8.1.0",
-		"typescript": "^3.0.1",
-		"typescript-eslint-parser": "^21.0.2"
-	}
+  "name": "@xcandy/ui-taro",
+  "version": "0.0.0-development",
+  "main": "dist/index.js",
+  "description": "A UI framework for taro.js",
+  "scripts": {
+    "release": "bash scripts/release.sh",
+    "test": "echo \"no test\"",
+    "lint": "npm run eslint",
+    "lint:fix": "npm run eslint:fix",
+    "eslint": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
+    "eslint:fix": "eslint --fix \"src/**/*.{js,jsx,ts,tsx}\"",
+    "ui:build": "cross-env TARO_BUILD_TYPE=ui taro build --ui",
+    "ui:ci": "TARO_BUILD_TYPE=ui taro build --ui",
+    "build:weapp": "taro build --type weapp",
+    "build:swan": "taro build --type swan",
+    "build:alipay": "taro build --type alipay",
+    "build:tt": "taro build --type tt",
+    "build:h5": "taro build --type h5",
+    "build:qq": "taro build --type qq",
+    "build:quickapp": "taro build --type quickapp",
+    "build:rn": "taro build --type rn",
+    "dev:weapp": "npm run build:weapp -- --watch",
+    "dev:swan": "npm run build:swan -- --watch",
+    "dev:alipay": "npm run build:alipay -- --watch",
+    "dev:tt": "npm run build:tt -- --watch",
+    "dev:h5": "npm run build:h5 -- --watch",
+    "dev:qq": "npm run build:qq -- --watch",
+    "dev:quickapp": "npm run build:quickapp -- --watch",
+    "dev:rn": "npm run build:rn -- --watch",
+    "update:site": "scripts/updateSite"
+  },
+  "author": "love@onelong.org",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/xxxcandy/xcandy-ui-taro"
+  },
+  "keywords": [
+    "taro",
+    "react",
+    "ui framework",
+    "weapp"
+  ],
+  "files": [
+    "dist",
+    "@types",
+    "src"
+  ],
+  "types": "./@types/index.d.ts",
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run lint && lint-staged",
+      "commit-msg": "commitlint -e $GIT_PARAMS"
+    }
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx}": [
+      "eslint --quiet --cache --fix",
+      "git add"
+    ]
+  },
+  "dependencies": {
+    "@tarojs/components": "1.3.5",
+    "@tarojs/components-qa": "^1.3.5",
+    "@tarojs/router": "1.3.5",
+    "@tarojs/taro": "1.3.5",
+    "@tarojs/taro-alipay": "1.3.5",
+    "@tarojs/taro-h5": "1.3.5",
+    "@tarojs/taro-quickapp": "^1.3.5",
+    "@tarojs/taro-swan": "1.3.5",
+    "@tarojs/taro-tt": "1.3.5",
+    "@tarojs/taro-weapp": "1.3.5",
+    "classnames": "^2.2.6",
+    "dayjs": "^1.8.0",
+    "nerv-devtools": "^1.4.3",
+    "nervjs": "^1.4.3"
+  },
+  "devDependencies": {
+    "@tarojs/plugin-babel": "1.3.5",
+    "@tarojs/plugin-csso": "1.3.5",
+    "@tarojs/plugin-sass": "1.3.5",
+    "@tarojs/plugin-uglifyjs": "1.3.5",
+    "@tarojs/webpack-runner": "1.3.5",
+    "@types/react": "^16.4.8",
+    "@types/webpack-env": "^1.13.6",
+    "@typescript-eslint/parser": "^1.9.0",
+    "babel-eslint": "^8.2.3",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-jsx-stylesheet": "^0.6.5",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "commitlint": "^8.0.0",
+    "cross-env": "^5.2.0",
+    "eslint": "^5.16.0",
+    "eslint-config-taro": "1.3.5",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jest": "^21.21.0",
+    "eslint-plugin-react": "^7.8.2",
+    "eslint-plugin-taro": "1.3.5",
+    "husky": "^1.3.1",
+    "lint-staged": "^8.1.0",
+    "stylelint": "9.3.0",
+    "stylelint-config-taro-rn": "1.3.0-beta.6",
+    "stylelint-taro-rn": "1.3.0-beta.6",
+    "typescript": "^3.0.1"
+  }
 }
