@@ -1,45 +1,60 @@
-import Taro, { Component } from '@tarojs/taro'
+import Taro, { useCallback, useState, useEffect } from '@tarojs/taro'
 import { View } from '@tarojs/components'
 import classnames from 'classnames'
 
 import './index.scss'
 
-type WantProps = {
-  onClickMask?: () => void
+const defaultProps = {
+  show: false,
+  maskBackgroundColor: '',
+  animationDuration: ''
 }
-type DefaultProps = {
+
+type Props = {
   show: boolean
-}
-type IProps = WantProps & DefaultProps
+  maskBackgroundColor?: string
+  onClickMask?: () => void
+  children?: any
+} & typeof defaultProps
 
-class XcMask extends Component<IProps> {
-  static defaultProps: DefaultProps = {
-    show: false
-  }
+const XcMask = (props: Props) => {
+  const { show, maskBackgroundColor, onClickMask, children } = props
+  const [maskClassnames, setMaskClassnames] = useState('xc-mask')
+  const [maskInterlayerClassnames, setMaskInterlayerClassnames] = useState('xc-mask__interlayer')
+  const [maskStyle, setMaskStyle] = useState<string | { [key: string]: any }>('')
 
-  handleMaskClick = () => {
-    const { onClickMask } = this.props
-    onClickMask && onClickMask()
-  }
-
-  render () {
-    const { show } = this.props
-    const maskClassNames = classnames('xc-mask', {
-      'xc-mask--show': show
+  useEffect(() => {
+    setMaskStyle({
+      'background-color': show ? maskBackgroundColor : ''
     })
-    const maskInterlayerClassNames = classnames('xc-mask__interlayer', {
-      'xc-mask__interlayer--show': show
-    })
-    return (
-      <View className={maskClassNames}>
-        {/* interlayer 用于触发对mask区域的点击，为了避免mask在关闭的动画过程中，用户再次点击仍然触发handleMaskClick */}
-        <View className={maskInterlayerClassNames} onClick={this.handleMaskClick} />
-        {this.props.children}
-      </View>
+  }, [maskBackgroundColor, show])
+
+  useEffect(() => {
+    setMaskClassnames(
+      classnames('xc-mask', {
+        'xc-mask--show': show
+      })
     )
-  }
+    setMaskInterlayerClassnames(
+      classnames('xc-mask__interlayer', {
+        'xc-mask__interlayer--show': show
+      })
+    )
+  }, [show])
+
+  const handleMaskClick = useCallback(() => {
+    onClickMask && onClickMask()
+  }, [onClickMask])
+
+  return (
+    <View className={maskClassnames} style={maskStyle}>
+      {/* interlayer 用于触发对mask区域的点击，为了避免mask在关闭的动画过程中，用户再次点击仍然触发handleMaskClick */}
+      <View className={maskInterlayerClassnames} onClick={handleMaskClick} />
+      {children}
+    </View>
+  )
 }
 
-export type Props = JSX.LibraryManagedAttributes<typeof XcMask, XcMask["props"]>
+XcMask.defaultProps = defaultProps
 
 export default XcMask
