@@ -1,4 +1,4 @@
-import Taro, { Component, Config } from '@tarojs/taro'
+import Taro, { Component } from '@tarojs/taro'
 import Index from './pages/index'
 
 import './app.scss'
@@ -8,7 +8,7 @@ if (process.env.NODE_ENV !== 'production' && process.env.TARO_ENV === 'h5') {
 }
 
 class App extends Component {
-  config: Config = {
+  config: Taro.Config = {
     pages: [
       'pages/index/index',
       'pages/layout/list/index',
