@@ -1,9 +1,8 @@
 import Taro, { Component } from '@tarojs/taro'
 import { View } from '@tarojs/components'
 
-import XcCard from '@xcComponents/Card'
+import Card from '@components/Card'
 import XcDrawer from '@xcComponents/Drawer'
-import XcMask from '@xcComponents/Mask'
 
 import './index.scss'
 
@@ -20,7 +19,7 @@ class DrawerPage extends Component {
     showDrawer3: false
   }
 
-  drawerSwitch1 = (placement?: string) => {
+  drawerSwitch1 = (placement?: 'top' | 'bottom') => {
     const { showDrawer1, placement1 } = this.state
     this.setState({
       showDrawer1: !showDrawer1,
@@ -28,7 +27,7 @@ class DrawerPage extends Component {
     })
   }
 
-  drawerSwitch2 = (placement?: string) => {
+  drawerSwitch2 = (placement?: 'left' | 'right') => {
     const { showDrawer2, placement2 } = this.state
     this.setState({
       showDrawer2: !showDrawer2,
@@ -43,11 +42,11 @@ class DrawerPage extends Component {
     })
   }
 
-  render () {
+  render() {
     const { showDrawer1, placement1, showDrawer2, placement2, showDrawer3 } = this.state
     return (
       <View className='page'>
-        <XcCard title='基础用法'>
+        <Card title='基础用法'>
           <View onClick={this.drawerSwitch1.bind(this, 'bottom')}>点击开启 - bottom</View>
           <View onClick={this.drawerSwitch1.bind(this, 'top')}>点击开启 - top</View>
 
@@ -55,10 +54,10 @@ class DrawerPage extends Component {
           <View onClick={this.drawerSwitch2.bind(this, 'left')}>点击开启 - left - 点击mask区域不能关闭</View>
 
           <View onClick={this.drawerSwitch3}>点击开启 - right - 没有mask区域</View>
-        </XcCard>
+        </Card>
 
         <XcDrawer show={showDrawer1} onClickMask={this.drawerSwitch1} placement={placement1}>
-          <View className='drawer-content__1'>
+          <View className='drawer-content drawer-content-1'>
             <View onClick={this.drawerSwitch1.bind(this, '')}>你可以点击这里关闭</View>
             <View>
               <View>content</View>
@@ -76,7 +75,7 @@ class DrawerPage extends Component {
         </XcDrawer>
 
         <XcDrawer show={showDrawer2} placement={placement2}>
-          <View className='drawer-content__2'>
+          <View className='drawer-content drawer-content-2'>
             <View onClick={this.drawerSwitch2.bind(this, '')}>你可以点击这里关闭</View>
             <View>
               <View>content</View>
@@ -93,8 +92,8 @@ class DrawerPage extends Component {
           </View>
         </XcDrawer>
 
-        <XcDrawer show={showDrawer3} placement='left' mask={false}>
-          <View className='drawer-content__3'>
+        <XcDrawer show={showDrawer3} placement='left' withoutMask>
+          <View className='drawer-content drawer-content-3'>
             <View onClick={this.drawerSwitch3}>你可以点击这里关闭</View>
             <View>
               <View>content</View>
